03-06 akşam 6.56 kodları

+++ server/controller/customerController.js
@@ -1,5 +1,6 @@
-import { response } from "express";
 import Customer from "../model/customerModel";
+import Receipt from "../model/receiptModel";
+import Park from "../model/parkModel";
 
 const index = (req, res) => {
 	res.json({ message: "Hello from Customers!" });
@@ -13,7 +14,10 @@ const addCustomer = async (req, res) => {
 			$or: [{ TCKN: TCKN }, { email: email }, { plate: plate }],
 		});
 		if (tempCustomer.length > 0) {
-			return res.json({ success: false, error: "Böyle bir müşteri zaten mevcut" });
+			return res.json({
+				success: false,
+				error: "Böyle bir müşteri zaten mevcut",
+			});
 		}
 
 		const newCustomer = await Customer.create({ fullName, TCKN, email, phone, plate });
@@ -31,7 +35,7 @@ const addCustomer = async (req, res) => {
 	}
 };
 
-const findCustomer = async (req, res) => {
+const findCustomerWithPlate = async (req, res) => {
 	// gelen plakayı al +
 	// plakayı veritabanında ara    +
 	// varsa müşteriyi döndür   +
@@ -59,10 +63,48 @@ const findCustomer = async (req, res) => {
 	}
 };
 
+const findCustomerWithId = async (req, res) => {
+	try {
+		const customerId = req.query.id;
+		const customer = await Customer.findById(customerId, {
+			updatedAt: 0,
+			__v: 0,
+		});
+
+		if (customer.length == 0) {
+			return res.json({
+				success: false,
+				error: "Böyle bir kullanıcı kayıtlı değil.",
+			});
+		}
+		// müşteri bulunmuştur artık döndürebiliriz.
+		return res.json({ success: true, customer: customer });
+	} catch (error) {
+		return res.json({ success: false, error: error.message });
+	}
+};
+
 const deleteCustomer = async (req, res) => {
 	console.log("deleteCustomer");
 	try {
 		const { customerId } = req.body;
+
+		const tempPark = await Park.find({
+			$and: [{ customer_id: customerId, park_state: 1 }],
+		});
+
+		const tempReceipt = await Receipt.find({
+			$and: [{ customer_id: customerId, receipt_state: 1 }],
+		});
+
+		if (tempPark.length > 0 || tempReceipt.length > 0) {
+			return res.json({
+				success: false,
+				message: "Bu müşterinin aktif park kaydı veya aktif fişi mevcut silinemez.",
+			});
+		}
+		
+		console.log(tempPark, tempReceipt);
 		const deletedUser = await Customer.findByIdAndDelete(customerId);
 
 		if (deletedUser) {
@@ -82,6 +124,36 @@ const deleteCustomer = async (req, res) => {
 	}
 };
 
+const updateCustomer = async (req, res) => {
+	try {
+		const customer = req.body;
+
+		const customerId = customer._id;
+		console.log(customer, customerId);
+
+		const updatedCustomer = await Customer.findByIdAndUpdate(customerId, customer, {
+			new: false,
+		});
+
+		if (!updatedCustomer) {
+			return res.json({
+				success: false,
+				message: "Güncelleme İşleminde Bir Hata Oluştu.",
+			});
+		}
+
+		return res.json({
+			success: true,
+			customer: updatedCustomer,
+			message: "Müşteri başarıyla güncellendi.",
+		});
+
+		return res.json({ success: true, message: "deneme mesaji." });
+	} catch (error) {
+		return res.json({ success: false, error: error.message });
+	}
+};
+
 const allCustomer = async (req, res) => {
 	try {
 		const customers = await Customer.find({}, {});
@@ -96,7 +168,10 @@ const activeCustomers = async (req, res) => {
 		const customers = await Customer.find({ customer_status: 1 });
 
 		if (customers.length == 0) {
-			return res.json({ success: false, error: "Aktif Müşteri Kaydı Bulunamadı." });
+			return res.json({
+				success: false,
+				error: "Aktif Müşteri Kaydı Bulunamadı.",
+			});
 		}
 
 		return res.json({ success: true, activeCustomers: customers });
@@ -110,7 +185,10 @@ const passiveCustomers = async (req, res) => {
 		const customers = await Customer.find({ customer_status: 0 });
 
 		if (customers.length == 0) {
-			return res.json({ success: false, error: "Pasif Müşteri Kaydı Bulunamadı." });
+			return res.json({
+				success: false,
+				error: "Pasif Müşteri Kaydı Bulunamadı.",
+			});
 		}
 
 		return res.json({ success: true, passiveCustomers: customers });
@@ -122,8 +200,10 @@ const passiveCustomers = async (req, res) => {
 export default {
 	index,
 	addCustomer,
-	findCustomer,
+	findCustomerWithPlate,
+	findCustomerWithId,
 	deleteCustomer,
+	updateCustomer,
 	allCustomer,
 	activeCustomers,
 	passiveCustomers,
